service-analysts/requirements.txt
boto3
mongoengine
PyJWT
pymongo
